Extract shared autoregressive decode runtime
