+++ Sources/MereRunCore/DecodeRuntimeExports.swift
@@ -1,0 +1,2 @@
+// Preserve source imports for callers of the shared decoding runtime.
+@_exported import MereRunDecode
