+++ Sources/MereRunCore/README.md
@@ -5,7 +5,8 @@ runtime families. Model metadata and installed lookup live in `MereRunModelKit`;
 Core re-exports those types for existing callers.
 Shared full-attention caches live in `MereRunKVCache`. Tensor loading, Qwen
 encoder layers, and FLUX.2/ZImage models live in `MereRunTensor`,
-`MereRunTextEncoder`, and `MereRunImageModels`. Core re-exports these libraries.
+`MereRunTextEncoder`, and `MereRunImageModels`. Core re-exports these libraries and `MereRunDecode`, which owns shared token
+sampling, decoding, streaming, and logprob diagnostics.
 
 - `Generation.swift`: image and chat request/response contracts.
 - `ImageGenerationOptions.swift` and `ImageGenerationPlan.swift`: typed image
